requirements.txt
requests==2.19.1
simplejson==3.16.0
six==1.11.0
pysha3==1.0.2
